+Adding support for dynamically updating the voxel grid based on the view and a view distance property +Adding optimization to the generation of the voxel blocks +Simplifying the voxel cell isosurface +minor fixes to the marching cubes calculation

+++ trunk/delta3d/inc/dtVoxel/marchingcubes.h
@@ -49,7 +49,7 @@ namespace dtVoxel
    0 will be returned if the grid cell is either totally above
    of totally below the isolevel.
    */
-   DT_VOXEL_EXPORT int PolygonizeCube(GRIDCELL g, double iso, TRIANGLE *tri);
+   DT_VOXEL_EXPORT int PolygonizeCube(GRIDCELL g, float iso, TRIANGLE *tri);
 
 
 } /* namespace dtVoxel */
